Переписал Entity так, чтобы он хранил параметры, котроые в таблице objects

+++ trainings/Maxim/src/main/java/com/annotation/Attribute.java
@@ -5,5 +5,5 @@ import java.lang.annotation.*;
 @Target(ElementType.FIELD)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface Attribute {
-    public int value();
+    public long value();
 }
